[Exchange] move ccxt connector to sub package
