+++ octobot_trading/exchanges/config/__init__.py
@@ -20,19 +20,11 @@ from octobot_trading.exchanges.config.exchange_config_data import (
     ExchangeConfig,
 )
 
-from octobot_trading.exchanges.config import exchange_settings_ccxt
-from octobot_trading.exchanges.config.exchange_settings_ccxt import (
-    CCXTExchangeConfig
-)
 from octobot_trading.exchanges.config import exchange_test_status
 from octobot_trading.exchanges.config.exchange_test_status import (
     ExchangeTestStatus
 )
 
-from octobot_trading.exchanges.config import ccxt_exchange_ui_settings
-from octobot_trading.exchanges.config.ccxt_exchange_ui_settings import (
-    initialize_experimental_exchange_settings,
-)
 from octobot_trading.exchanges.config import backtesting_exchange_config
 from octobot_trading.exchanges.config.backtesting_exchange_config import (
     BacktestingExchangeConfig,
@@ -41,7 +33,5 @@ from octobot_trading.exchanges.config.backtesting_exchange_config import (
 __all__ = [
     "ExchangeConfig",
     "BacktestingExchangeConfig",
-    "CCXTExchangeConfig",
     "ExchangeTestStatus",
-    "initialize_experimental_exchange_settings",
 ]
