+++ octobot_trading/exchanges/connectors/ccxt/__init__.py
@@ -14,31 +14,31 @@
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
 
-from octobot_trading.exchanges.connectors import exchange_simulator
-from octobot_trading.exchanges.connectors.exchange_simulator import (
-    ExchangeSimulator,
-)
 from octobot_trading.exchanges.connectors.ccxt import ccxt_exchange
 from octobot_trading.exchanges.connectors.ccxt.ccxt_exchange import (
     CCXTExchange,
-)
-from octobot_trading.exchanges.connectors import abstract_websocket_connector
-from octobot_trading.exchanges.connectors.abstract_websocket_connector import (
-    AbstractWebsocketConnector,
 )
 from octobot_trading.exchanges.connectors.ccxt import ccxt_websocket_connector
 from octobot_trading.exchanges.connectors.ccxt.ccxt_websocket_connector import (
     CCXTWebsocketConnector,
 )
-from octobot_trading.exchanges.connectors import cryptofeed_websocket_connector
-from octobot_trading.exchanges.connectors.cryptofeed_websocket_connector import (
-    CryptofeedWebsocketConnector,
+from octobot_trading.exchanges.connectors.ccxt import ccxt_exchange_ui_settings
+from octobot_trading.exchanges.connectors.ccxt.ccxt_exchange_ui_settings import (
+    initialize_experimental_exchange_settings,
+)
+from octobot_trading.exchanges.connectors.ccxt import exchange_settings_ccxt
+from octobot_trading.exchanges.connectors.ccxt.exchange_settings_ccxt import (
+    CCXTExchangeConfig,
+)
+from octobot_trading.exchanges.connectors.ccxt import exchange_settings_ccxt_generic
+from octobot_trading.exchanges.connectors.ccxt.exchange_settings_ccxt_generic import (
+    GenericCCXTExchangeConfig,
 )
 
 __all__ = [
-    "ExchangeSimulator",
     "CCXTExchange",
-    "AbstractWebsocketConnector",
     "CCXTWebsocketConnector",
-    "CryptofeedWebsocketConnector",
+    "initialize_experimental_exchange_settings",
+    "CCXTExchangeConfig",
+    "GenericCCXTExchangeConfig",
 ]
