+++ octobot_trading/exchanges/connectors/ccxt/ccxt_exchange.py
@@ -32,7 +32,7 @@ import octobot_trading.enums as enums
 import octobot_trading.errors
 import octobot_trading.exchanges as exchanges
 import octobot_trading.exchanges.abstract_exchange as abstract_exchange
-import octobot_trading.exchanges.config.exchange_settings_ccxt as exchange_settings_ccxt
+import octobot_trading.exchanges.connectors.ccxt.exchange_settings_ccxt as exchange_settings_ccxt
 import octobot_trading.personal_data as personal_data
 from octobot_trading.enums import ExchangeConstantsOrderColumns as ecoc, OrderStatus, PositionSubType
