+++ octobot_trading/exchanges/connectors/ccxt/exchange_settings_ccxt.pxd
