+++ octobot_trading/exchanges/connectors/exchange_settings.py
