+++ octobot_trading/exchanges/implementations/ccxt_exchange_commons.py
@@ -19,9 +19,9 @@ import typing
 
 import octobot_commons.enums as common_enums
 import octobot_trading.enums as enums
-import octobot_trading.exchanges.config.exchange_settings_ccxt_generic as exchange_settings_ccxt_generic
-import octobot_trading.exchanges.config.exchange_settings as exchange_settings
-import octobot_trading.exchanges.config.ccxt_exchange_ui_settings as ccxt_exchange_ui_settings
+import octobot_trading.exchanges.connectors.ccxt.exchange_settings_ccxt_generic as exchange_settings_ccxt_generic
+import octobot_trading.exchanges.connectors.exchange_settings as exchange_settings
+import octobot_trading.exchanges.connectors.ccxt.ccxt_exchange_ui_settings as ccxt_exchange_ui_settings
 import octobot_trading.exchanges.types as exchanges_types
 import octobot_trading.exchanges.connectors as exchange_connectors
